Factoring out some parts out of rdfquery
